Re-structure of functions and cleanup
Main changes:

 - moved some functions from src/cuda_oddeven_generic.cu to other files
   for readability
 - added some if statements to apply CUDA-Schwarz at depth=0 only
 - some small rearrangements of data in structs in src/main_pre_def_generic.h
 - cleanup of src/schwarz_generic.c
 - re-organized src/schwarz_generic.c to have a better split of initial
   operations according to: initialize, alloc, setup, free
