+++ src/schwarz_generic.c
@@ -35,30 +35,11 @@ void smoother_PRECISION_def( level_struct *l ) {
 }
 
 
-#ifdef CUDA_OPT
-void smoother_PRECISION_def_CUDA( level_struct *l ) {
-
-  if ( g.method >= 0 )
-    schwarz_PRECISION_def_CUDA( &(l->s_PRECISION), &(g.op_double), l );
-}
-#endif
-
-
 void smoother_PRECISION_free( level_struct *l ) {
   
   if ( g.method >= 0 )
     schwarz_PRECISION_free( &(l->s_PRECISION), l );
 }
-
-
-#ifdef CUDA_OPT
-void smoother_PRECISION_free_CUDA( level_struct *l ) {
-  
-  if( l->depth==0 && g.odd_even ){
-    schwarz_PRECISION_free_CUDA( &(l->s_PRECISION), l );
-  }
-}
-#endif
 
 
 void schwarz_PRECISION_init( schwarz_PRECISION_struct *s, level_struct *l ) {
@@ -88,15 +69,6 @@ void schwarz_PRECISION_init( schwarz_PRECISION_struct *s, level_struct *l ) {
   s->block_list_length = NULL;
   s->num_colors = 0;
 }
-
-
-#ifdef CUDA_OPT
-void schwarz_PRECISION_init_CUDA( schwarz_PRECISION_struct *s, level_struct *l ) {
-
-  // TODO !
-
-}
-#endif
 
 
 void schwarz_PRECISION_alloc( schwarz_PRECISION_struct *s, level_struct *l ) {
@@ -250,56 +222,6 @@ void schwarz_PRECISION_alloc( schwarz_PRECISION_struct *s, level_struct *l ) {
 }
 
 
-// FIXME
-#ifdef CUDA_OPT
-void schwarz_PRECISION_alloc_CUDA( schwarz_PRECISION_struct *s, level_struct *l ) {
-
-  int vs, i;
-
-  s->streams = (cudaStream_t*) malloc( 1 * sizeof(cudaStream_t) );
-  cuda_safe_call( cudaStreamCreate( &(s->streams[0]) ) );
-
-  vs = (l->depth==0)?l->inner_vector_size:l->vector_size;
-
-  // FIXME: what is the role of 'vs' here ?
-  //MALLOC( s->buf1, complex_PRECISION, vs+3*l->schwarz_vector_size );
-  // Buffers for the computations in SAP
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf1 )), vs*sizeof(cu_cmplx_PRECISION) ) );
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf2 )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf3 )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf4 )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-
-  // FIXME: the size of the following buffer is wrong
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf5 )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-  //cuda_vector_PRECISION eta_dev;
-  //cuda_safe_call( cudaMallocHost( (void**)&(eta_dev), (nb_thread_end - nb_thread_start)*s->block_vector_size * sizeof(complex_PRECISION) ) );
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).buf6 )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-
-  if ( l->depth == 0 ) {
-    // TODO: is l->inner_vector_size the ACTUAL size in this allocation ?
-    for( i=0; i<4; i++ ){
-      cuda_safe_call( cudaMalloc( (void**)&( (s->s_on_gpu_cpubuff).oe_buf[i] ), l->inner_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-    }
-  }
-
-  cuda_safe_call( cudaMalloc( (void**) (&( s->s_on_gpu )), 1*sizeof(schwarz_PRECISION_struct_on_gpu) ) );
-
-  // TODO: do this at all MG levels ?
-  if( l->depth==0 ){
-    // these buffers are introduced to make local_minres_PRECISION thread-safe
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).local_minres_buffer[0] )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).local_minres_buffer[1] )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).local_minres_buffer[2] )), l->schwarz_vector_size*sizeof(cu_cmplx_PRECISION) ) );
-
-    //MALLOC( s->local_minres_buffer[0], complex_PRECISION, l->schwarz_vector_size );
-    //MALLOC( s->local_minres_buffer[1], complex_PRECISION, l->schwarz_vector_size );
-    //MALLOC( s->local_minres_buffer[2], complex_PRECISION, l->schwarz_vector_size );
-  }
-
-}
-#endif
-
-
 void schwarz_PRECISION_free( schwarz_PRECISION_struct *s, level_struct *l ) {
   
   int i, n, mu, nu, *bl = l->block_lattice, vs = (l->depth==0)?l->inner_vector_size:l->vector_size;
@@ -359,11 +281,13 @@ void schwarz_PRECISION_free( schwarz_PRECISION_struct *s, level_struct *l ) {
   s->oe_bbuf[2] = NULL; s->oe_bbuf[3] = NULL; s->oe_bbuf[4] = NULL; s->oe_bbuf[5] = NULL;
 
   // FIXME: is pinned memory still to be used ?
-#ifdef CUDA_OPT  
-  cuda_safe_call( cudaFreeHost(s->buf1) );
-#else
+  //#ifdef CUDA_OPT  
+    //cuda_safe_call( cudaFreeHost(s->buf1) );
+  //#else
+    //FREE( s->buf1, complex_PRECISION, vs+3*l->schwarz_vector_size );
+  //#endif
+
   FREE( s->buf1, complex_PRECISION, vs+3*l->schwarz_vector_size );
-#endif
   s->buf2 = NULL; s->buf3 = NULL;
   s->buf4 = NULL;
   
@@ -398,31 +322,6 @@ void schwarz_PRECISION_free( schwarz_PRECISION_struct *s, level_struct *l ) {
   }
 #endif
 }
-
-
-// TODO
-#ifdef CUDA_OPT
-void schwarz_PRECISION_free_CUDA( schwarz_PRECISION_struct *s, level_struct *l ) {
-
-  int i;
-
-  free( s->streams );
-
-  cuda_safe_call( cudaFree( (s->cu_s).buf1 ) );
-  cuda_safe_call( cudaFree( (s->cu_s).buf2 ) );
-  cuda_safe_call( cudaFree( (s->cu_s).buf3 ) );
-  cuda_safe_call( cudaFree( (s->cu_s).buf4 ) );
-
-  if ( l->depth == 0 ) {
-    // TODO: is l->inner_vector_size the ACTUAL size in this allocation ?
-    for( i=0; i<4; i++ ){
-      cuda_safe_call( cudaFree( (s->s_on_gpu_cpubuff).oe_buf[i] ) );
-    }
-  }
-
-  cuda_safe_call( cudaFree( s->s_on_gpu ) );
-}
-#endif
 
 
 void schwarz_layout_PRECISION_define( schwarz_PRECISION_struct *s, level_struct *l ) {
@@ -1191,800 +1090,6 @@ void schwarz_PRECISION_setup( schwarz_PRECISION_struct *s, operator_double_struc
 #endif
 
 
-// TODO
-#ifdef CUDA_OPT
-void schwarz_PRECISION_setup_CUDA( schwarz_PRECISION_struct *s, operator_double_struct *op_in, level_struct *l ) {
-
-  int color, comms_ctr, noncomms_ctr, color_ctr, i, j, k;
-
-  // FIXME: call malloc with error check
-
-  // TODO: should I move some of these lines to another schwarz_* function ?
-
-  s->nr_DD_blocks_in_comms = (int*) malloc( s->num_colors*sizeof(int) );
-  s->nr_DD_blocks_notin_comms = (int*) malloc( s->num_colors*sizeof(int) );
-
-  s->nr_DD_blocks = (int*) malloc( s->num_colors*sizeof(int) );
-
-  s->DD_blocks_in_comms = (int**) malloc( s->num_colors*sizeof(int*) );
-  s->DD_blocks_notin_comms = (int**) malloc( s->num_colors*sizeof(int*) );
-
-  s->DD_blocks = (int**) malloc( s->num_colors*sizeof(int*) );
-
-  //printf("s->num_colors = %d\n", s->num_colors);
-
-  for(color=0; color<s->num_colors; color++){
-    s->nr_DD_blocks_notin_comms[color] = 0;
-    s->nr_DD_blocks_in_comms[color] = 0;
-    s->nr_DD_blocks[color] = 0;
-    for(i=0; i<s->num_blocks; i++){
-      if ( color == s->block[i].color && s->block[i].no_comm ) {
-        s->nr_DD_blocks_notin_comms[color]++;
-      }
-      else if( color == s->block[i].color && !s->block[i].no_comm ){
-        s->nr_DD_blocks_in_comms[color]++;
-      }
-      if( color == s->block[i].color ){
-        s->nr_DD_blocks[color]++;
-      }
-    }
-  }
-
-  for(color=0; color<s->num_colors; color++){
-    //printf("s->nr_DD_blocks_in_comms[color] = %d\n", s->nr_DD_blocks_in_comms[color]);
-    //printf("s->nr_DD_blocks_notin_comms[color] = %d\n", s->nr_DD_blocks_notin_comms[color]);
-    s->DD_blocks_in_comms[color] = (int*) malloc( s->nr_DD_blocks_in_comms[color]*sizeof(int) );
-    s->DD_blocks_notin_comms[color] = (int*) malloc( s->nr_DD_blocks_notin_comms[color]*sizeof(int) );
-    s->DD_blocks[color] = (int*) malloc( s->nr_DD_blocks[color]*sizeof(int) );
-  }
-
-  for(color=0; color<s->num_colors; color++){
-    comms_ctr = 0;
-    noncomms_ctr = 0;
-    color_ctr = 0;
-    for(i=0; i<s->num_blocks; i++){
-      if ( color == s->block[i].color && s->block[i].no_comm ) {
-        s->DD_blocks_notin_comms[color][noncomms_ctr] = i;
-        noncomms_ctr++;
-      }
-      else if( color == s->block[i].color && !s->block[i].no_comm ){
-        s->DD_blocks_in_comms[color][comms_ctr] = i;
-        comms_ctr++;
-      }
-      if( color == s->block[i].color ){
-        s->DD_blocks[color][color_ctr] = i;
-        color_ctr++;
-      }
-    }
-  }
-
-  (s->cu_s).DD_blocks_in_comms = (int**) malloc( s->num_colors*sizeof(int*) );
-  (s->cu_s).DD_blocks_notin_comms = (int**) malloc( s->num_colors*sizeof(int*) );
-
-  (s->cu_s).DD_blocks = (int**) malloc( s->num_colors*sizeof(int*) );
-
-  for(color=0; color<s->num_colors; color++){
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).DD_blocks_in_comms[color] )), s->nr_DD_blocks_in_comms[color]*sizeof(int) ) );
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).DD_blocks_notin_comms[color] )), s->nr_DD_blocks_notin_comms[color]*sizeof(int) ) );
-
-    cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).DD_blocks[color] )), s->nr_DD_blocks[color]*sizeof(int) ) );
-  }
-
-  for(color=0; color<s->num_colors; color++){
-    cuda_safe_call( cudaMemcpy((s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color], s->nr_DD_blocks_in_comms[color]*sizeof(int), cudaMemcpyHostToDevice) );
-    cuda_safe_call( cudaMemcpy((s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color], s->nr_DD_blocks_notin_comms[color]*sizeof(int), cudaMemcpyHostToDevice) );
-
-    cuda_safe_call( cudaMemcpy((s->cu_s).DD_blocks[color], s->DD_blocks[color], s->nr_DD_blocks[color]*sizeof(int), cudaMemcpyHostToDevice) );
-  }
-
-
-
-
-
-
-
-
-  //s->block_boundary_length[8] = n;
-
-  int *block_arr_buff;
-
-  cuda_safe_call( cudaMalloc( (void**) (&( block_arr_buff )), s->block_boundary_length[8] * s->num_blocks * sizeof(int) ) );
-
-  for ( i=0; i<s->num_blocks; i++ ) {
-    s->block[i].bt_on_gpu = block_arr_buff + i * s->block_boundary_length[8];
-  }
-
-  //printf("s->block_boundary_length[8]=%d\n", s->block_boundary_length[8]);
-
-  //MALLOC( s->block[i].bt_on_gpu, int, s->block_boundary_length[8] * s->num_blocks );
-
-  cuda_safe_call( cudaMalloc( (void**) (&( (s->cu_s).block )), s->num_blocks*sizeof(block_struct) ) );
-  cuda_safe_call( cudaMemcpy((s->cu_s).block, s->block, s->num_blocks*sizeof(block_struct), cudaMemcpyHostToDevice) );
-
-  for ( i=0; i<s->num_blocks; i++ ) {
-    //s->block[i].bt_on_gpu = block_arr_buff + i * s->block_boundary_length[8];
-    cuda_safe_call( cudaMemcpy( s->block[i].bt_on_gpu, s->block[i].bt, s->block_boundary_length[8] * sizeof(int), cudaMemcpyHostToDevice ) );
-  }
-
-
-  for( i=0; i<9; i++ ){
-    (s->s_on_gpu_cpubuff).block_boundary_length[i] = s->block_boundary_length[i];
-    //(s->s_on_gpu_cpubuff).dir_length_odd[i] = s->dir_length_odd[i];
-  }
-
-
-
-
-
-
-  // CRITICAL DATA MOVEMENT: copying op.oe_clover_vectorized to the GPU
-  //-------------------------------------------------------------------------------------------------------------------------
-  int b, h, nr_DD_sites;
-
-  nr_DD_sites = s->num_block_sites;
-
-  schwarz_PRECISION_struct_on_gpu *out = &(s->s_on_gpu_cpubuff);
-  schwarz_PRECISION_struct *in = s;
-
-  cu_cmplx_PRECISION *buf_D_oe_cpu, *buf_D_oe_cpu_bare;
-  cu_config_PRECISION *buf_D_oe_gpu;
-
-  if(g.csw != 0){
-    buf_D_oe_cpu = (cu_cmplx_PRECISION*) malloc( 72 * nr_DD_sites*in->num_blocks * sizeof(cu_cmplx_PRECISION) );
-    cuda_safe_call( cudaMalloc( (void**) &(buf_D_oe_gpu), 72 * sizeof(cu_config_PRECISION) * nr_DD_sites*in->num_blocks ) );
-  }
-  else{
-    buf_D_oe_cpu = (cu_cmplx_PRECISION*) malloc( 12 * nr_DD_sites*in->num_blocks * sizeof(cu_cmplx_PRECISION) );
-    cuda_safe_call( cudaMalloc( (void**) &(buf_D_oe_gpu), 12 * sizeof(cu_config_PRECISION) * nr_DD_sites*in->num_blocks ) );
-  }
-  buf_D_oe_cpu_bare = buf_D_oe_cpu;
-
-  if(g.csw != 0){
-
-    PRECISION *op_oe_vect_bare = in->op.oe_clover_vectorized;
-    PRECISION *op_oe_vect = op_oe_vect_bare;
-
-    PRECISION M_tmp[144];
-    PRECISION *M_tmp1, *M_tmp2;
-    M_tmp1 = M_tmp;
-    M_tmp2 = M_tmp + 72;
-
-    for(b=0; b < in->num_blocks; b++){
-      for(h=0; h<nr_DD_sites; h++){
-        //the following snippet of code was taken from function sse_site_clover_invert_float(...) in sse_dirac.c
-        for ( k=0; k<12; k+=SIMD_LENGTH_float ) {
-          for ( j=0; j<6; j++ ) {
-            for ( i=k; i<k+SIMD_LENGTH_float; i++ ) {
-              if ( i<6 ) {
-                M_tmp1[12*j+i] = *op_oe_vect;
-                M_tmp1[12*j+i+6] = *(op_oe_vect+SIMD_LENGTH_float);
-              } else {
-                M_tmp2[12*j+i-6] = *op_oe_vect;
-                M_tmp2[12*j+i] = *(op_oe_vect+SIMD_LENGTH_float);
-              }
-              op_oe_vect++;
-            }
-            op_oe_vect += SIMD_LENGTH_float;
-          }
-        }
-
-        //the following snippet of code was taken from the function sse_cgem_inverse(...) within sse_blas_vectorized.h
-        int N=6;
-        //cu_cmplx_PRECISION tmpA[2*N*N];
-        cu_cmplx_PRECISION *tmpA_1, *tmpA_2;
-        tmpA_1 = buf_D_oe_cpu;
-        tmpA_2 = buf_D_oe_cpu + N*N;
-        for ( j=0; j<N; j++ ) {
-          for ( i=0; i<N; i++ ) {
-            tmpA_1[i+N*j] = make_cu_cmplx_PRECISION(M_tmp1[2*j*N+i], M_tmp1[(2*j+1)*N+i]);
-            tmpA_2[i+N*j] = make_cu_cmplx_PRECISION(M_tmp2[2*j*N+i], M_tmp2[(2*j+1)*N+i]);
-            //printf("%f + i%f\n", cu_creal_PRECISION(tmpA_2[i+N*j]), cu_cimag_PRECISION(tmpA_2[i+N*j]));
-          }
-        }
-
-        /*
-        if( b==100 && h==100 ){
-        for( i=0; i<N; i++ ){
-          for( j=0; j<N; j++ ){
-            if( g.my_rank ==0 ) printf(" %f+i%f |", cu_creal_PRECISION(buf_D_oe_cpu[i+j*N]), cu_cimag_PRECISION(buf_D_oe_cpu[i+j*N]));
-          }
-          if( g.my_rank==0 ) printf("\n");
-        }
-
-        printf("\n");
-        printf("\n");
-
-        if( b==100 && h==100 ){
-        for( i=0; i<N; i++ ){
-          for( j=0; j<N; j++ ){
-            if( g.my_rank ==0 ) printf(" %f+i%f |", cu_creal_PRECISION((buf_D_oe_cpu+36)[i+j*N]), cu_cimag_PRECISION((buf_D_oe_cpu+36)[i+j*N]));
-          }
-          if( g.my_rank==0 ) printf("\n");
-        }
-        }
-
-        printf("\n");
-        printf("\n");
-        printf("\n");
-        //MPI_Barrier(MPI_COMM_WORLD);
-        //MPI_Finalize();
-        //exit(0);
-        }
-        */
-
-        buf_D_oe_cpu += 72;
-        op_oe_vect_bare += 144;
-        op_oe_vect = op_oe_vect_bare;
-      }
-    }
-  }
-  else{
-    //TODO !!
-    //vector_PRECISION_copy( op->oe_clover, op->clover, 0, l->inner_vector_size, l );
-  }
-
-
-  // making use of Doo^-1 (i.e. buf_D_oe_cpu_bare) being Hermitian to store in reduced form
-  // IMPORTANT: this matrix is stored in column-form
-
-  cu_cmplx_PRECISION *buf_D_oe_cpu_gpustorg=0, *buf_D_oe_gpu_gpustorg=0, *buf_D_oe_cpu_gpustorg_bare=0;
-
-  if(g.csw != 0){
-    buf_D_oe_cpu_gpustorg = (cu_cmplx_PRECISION*) malloc( 42 * nr_DD_sites*in->num_blocks * sizeof(cu_cmplx_PRECISION) );
-    cuda_safe_call( cudaMalloc( (void**) &(buf_D_oe_gpu_gpustorg), 42 * sizeof(cu_cmplx_PRECISION) * nr_DD_sites*in->num_blocks ) );
-  }
-  else{
-    // TODO !
-    //buf_D_oe_cpu = (cu_cmplx_PRECISION*) malloc( 12 * nr_DD_sites*in->num_blocks * sizeof(cu_cmplx_PRECISION) );
-    //cuda_safe_call( cudaMalloc( (void**) &(buf_D_oe_gpu), 12 * sizeof(cu_config_PRECISION) * nr_DD_sites*in->num_blocks ) );
-  }
-
-  buf_D_oe_cpu = buf_D_oe_cpu_bare;
-  buf_D_oe_cpu_gpustorg_bare = buf_D_oe_cpu_gpustorg;
-
-  for(b=0; b < in->num_blocks; b++){
-    for(h=0; h<nr_DD_sites; h++){
-      int N=6, k=0;
-      for ( j=0; j<N; j++ ) {
-        for ( i=j; i<N; i++ ) {
-
-          (buf_D_oe_cpu_gpustorg +    0)[k] = (buf_D_oe_cpu +    0)[i+j*N];
-          (buf_D_oe_cpu_gpustorg + 42/2)[k] = (buf_D_oe_cpu + 72/2)[i+j*N];
-          k++;
-
-        }
-      }
-
-
-        /*
-        int matrx_indx, local_idx;
-        if( b==100 && h==100 ){
-        for( i=0; i<2; i++ ){
-          for( local_idx=0; local_idx<N; local_idx++ ){
-            for( j=0; j<N; j++ ){
-
-              if( local_idx>j ){
-                matrx_indx = 21 - (5-j+1)*((5-j+1)+1)/2 + (local_idx-j);
-              }
-              else if( local_idx==j ){
-                matrx_indx = 21 - (5-j+1)*((5-j+1)+1)/2;
-              }
-              else{
-                matrx_indx = 21 - (5-local_idx+1)*((5-local_idx+1)+1)/2 + (j-local_idx);
-              }
-
-              //if( (threadIdx.x + blockDim.x * blockIdx.x)==0 ){
-              //  printf("local_idx=%d, j=%d\n", local_idx, j);
-              //}
-
-              if( local_idx>j || local_idx==j ){
-                //eta_site[ 6*i + local_idx ] = cu_cadd_PRECISION( eta_site[ 6*i + local_idx ], cu_cmul_PRECISION( (op_clov_vect_site + i*21)[matrx_indx], phi_site[j + 6*i] ) );
-                if( g.my_rank ==0 ) printf(" %f+i%f |", cu_creal_PRECISION((buf_D_oe_cpu_gpustorg+i*21)[matrx_indx]), cu_cimag_PRECISION((buf_D_oe_cpu_gpustorg+i*21)[matrx_indx]));
-              }
-              else{
-                //eta_site[ 6*i + local_idx ] = cu_cadd_PRECISION( eta_site[ 6*i + local_idx ], cu_cmul_PRECISION( cu_conj_PRECISION((op_clov_vect_site + i*21)[matrx_indx]), phi_site[j + 6*i] ) );
-                if( g.my_rank ==0 ) printf(" %f+i%f |", cu_creal_PRECISION(cu_conj_PRECISION((buf_D_oe_cpu_gpustorg+i*21)[matrx_indx])), cu_cimag_PRECISION(cu_conj_PRECISION((buf_D_oe_cpu_gpustorg+i*21)[matrx_indx])));
-              }
-
-              //if( g.my_rank ==0 ) printf(" %f+i%f |", cu_creal_PRECISION(buf_D_oe_cpu[i*N+j]), cu_cimag_PRECISION(buf_D_oe_cpu[i*N+j]));
-
-            }
-            if( g.my_rank==0 ) printf("\n");
-          }
-          if( g.my_rank==0 ) printf("\n\n");
-        }
-        }
-
-        //printf("\n");
-        //MPI_Barrier(MPI_COMM_WORLD);
-        //MPI_Finalize();
-        //exit(0);
-        */
-
-
-      buf_D_oe_cpu += 72;
-      buf_D_oe_cpu_gpustorg += 42;
-    }
-  }
-
-  //MPI_Barrier(MPI_COMM_WORLD);
-  //MPI_Abort(MPI_COMM_WORLD, 911);
-
-  if(g.csw != 0){
-    cuda_safe_call( cudaMemcpy(buf_D_oe_gpu, buf_D_oe_cpu_bare, 72*sizeof(cu_cmplx_PRECISION)*nr_DD_sites*in->num_blocks, cudaMemcpyHostToDevice) );
-  }
-  else{
-    cuda_safe_call( cudaMemcpy(buf_D_oe_gpu, buf_D_oe_cpu_bare, 12*sizeof(cu_cmplx_PRECISION)*nr_DD_sites*in->num_blocks, cudaMemcpyHostToDevice) );
-  }
-
-  if(g.csw != 0){
-    cuda_safe_call( cudaMemcpy(buf_D_oe_gpu_gpustorg, buf_D_oe_cpu_gpustorg_bare, 42*sizeof(cu_cmplx_PRECISION)*nr_DD_sites*in->num_blocks, cudaMemcpyHostToDevice) );
-  }
-  else{
-    // TODO !
-    //cuda_safe_call( cudaMemcpy(buf_D_oe_gpu, buf_D_oe_cpu_bare, 12*sizeof(cu_cmplx_PRECISION)*nr_DD_sites*in->num_blocks, cudaMemcpyHostToDevice) );
-    buf_D_oe_gpu_gpustorg = NULL;
-    buf_D_oe_cpu_gpustorg_bare = NULL;
-  }
-
-  free(buf_D_oe_cpu_bare);
-  free(buf_D_oe_cpu_gpustorg_bare);
-
-  //cudaMemcpy(&(out->op.oe_clover_vectorized), &buf_D_oe_gpu, sizeof(cu_config_PRECISION*), cudaMemcpyHostToDevice);
-  out->op.oe_clover_vectorized = buf_D_oe_gpu;
-  out->op.oe_clover_gpustorg = buf_D_oe_gpu_gpustorg;
-  //-------------------------------------------------------------------------------------------------------------------------
-
-
-
-
-
-  //-------------------------------------------------------------------------------------------------------------------------
-
-  int n=l->num_inner_lattice_sites;
-
-  // less memory required if csw==0
-
-  if(g.csw != 0){
-    cuda_safe_call( cudaMalloc( (void**) &(out->op.clover_gpustorg), 42 * sizeof(cu_cmplx_PRECISION) * n ) );
-    cuda_safe_call( cudaMemcpy( out->op.clover_gpustorg, s->op.clover, 42 * sizeof(cu_cmplx_PRECISION) * n, cudaMemcpyHostToDevice ) );
-  }
-  else{
-    cuda_safe_call( cudaMalloc( (void**) &(out->op.clover_gpustorg), 12 * sizeof(cu_cmplx_PRECISION) * n ) );
-    cuda_safe_call( cudaMemcpy( out->op.clover_gpustorg, s->op.clover, 12 * sizeof(cu_cmplx_PRECISION) * n, cudaMemcpyHostToDevice ) );
-  }
-
-
-  //-------------------------------------------------------------------------------------------------------------------------
-
-
-
-
-
-
-  int dir, nr_oe_elems_hopp;
-  for( dir=0; dir<4; dir++ ){
-    (s->s_on_gpu_cpubuff).dir_length_even[dir] = s->dir_length_even[dir];
-    (s->s_on_gpu_cpubuff).dir_length_odd[dir] = s->dir_length_odd[dir];
-  }
-  for( dir=0; dir<4; dir++ ){
-    nr_oe_elems_hopp = s->dir_length_even[dir] + s->dir_length_odd[dir];
-    cuda_safe_call( cudaMalloc( (void**)&( (s->s_on_gpu_cpubuff).oe_index[dir] ), nr_oe_elems_hopp*sizeof(int) ) );
-    cuda_safe_call( cudaMemcpy( (s->s_on_gpu_cpubuff).oe_index[dir], s->oe_index[dir], nr_oe_elems_hopp*sizeof(int), cudaMemcpyHostToDevice ) );
-  }
-
-  int nr_elems_block_d_plus;
-  for( dir=0; dir<4; dir++ ){
-    (s->s_on_gpu_cpubuff).dir_length[dir] = s->dir_length[dir];
-    (s->s_on_gpu_cpubuff).dir_length[dir] = s->dir_length[dir];
-  }
-  for( dir=0; dir<4; dir++ ){
-    nr_elems_block_d_plus = s->dir_length[dir];
-    cuda_safe_call( cudaMalloc( (void**)&( (s->s_on_gpu_cpubuff).index[dir] ), nr_elems_block_d_plus*sizeof(int) ) );
-    cuda_safe_call( cudaMemcpy( (s->s_on_gpu_cpubuff).index[dir], s->index[dir], nr_elems_block_d_plus*sizeof(int), cudaMemcpyHostToDevice ) );
-  }
-
-  cuda_safe_call( cudaMalloc( (void**)&( (s->s_on_gpu_cpubuff).op.neighbor_table ), (l->depth==0?4:5)*l->num_inner_lattice_sites*sizeof(int) ) );
-  cuda_safe_call( cudaMemcpy( (s->s_on_gpu_cpubuff).op.neighbor_table, s->op.neighbor_table, (l->depth==0?4:5)*l->num_inner_lattice_sites*sizeof(int), cudaMemcpyHostToDevice ) );
-
-  int type=_SCHWARZ, nls, coupling_site_size, nr_elems_D;
-  if ( l->depth == 0 ) {
-    coupling_site_size = 4*9;
-  } else {
-    coupling_site_size = 4*l->num_lattice_site_var*l->num_lattice_site_var;
-  }
-
-  nls = (type==_ORDINARY)?l->num_inner_lattice_sites:2*l->num_lattice_sites-l->num_inner_lattice_sites;
-  nr_elems_D = coupling_site_size*nls;
-
-  cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).op.D ), nr_elems_D*sizeof(cu_config_PRECISION) ) );
-  cuda_safe_call( cudaMemcpy((s->s_on_gpu_cpubuff).op.D, (cu_config_PRECISION*)(s->op.D), nr_elems_D*sizeof(cu_config_PRECISION), cudaMemcpyHostToDevice) );
-
-
-  //----------------------------------------------------------------
-
-  // transforming s->op.D to optimal form for accesses from the GPU
-
-  // 0, 1, 2, 3 ----> plus-even, plus-odd, minus-even, minus-odd
-  cu_cmplx_PRECISION *Dgpu[16];
-  int nr_elems_Dgpu[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
-  cu_cmplx_PRECISION *Dgpu_buff;
-  //int amount;
-  int a1, n1, a2, n2;
-  int DD_start;
-
-  int *ind;
-  int **index = s->oe_index;
-
-  //int num_latt_site_var = ;
-
-  config_PRECISION Dbuff;
-  config_PRECISION D_pt;
-
-  // do PLUS first
-
-  // ... for plus, do EVEN
-
-  //amount = _EVEN_SITES;
-
-  for( dir=0; dir<4; dir++ ){
-
-    a1 = 0; n1 = s->dir_length_even[dir];
-
-    nr_elems_Dgpu[0*8 + 0*4 + dir] =  (n1-a1);
-    nr_elems_Dgpu[0*8 + 0*4 + dir] *= s->num_blocks;
-
-    // 0*8 + 0*4 + dir ---> PLUS(0), EVEN(0), dir
-    Dgpu[0*8 + 0*4 + dir] = (cu_cmplx_PRECISION*) malloc( nr_elems_Dgpu[0*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) );
-
-    Dgpu_buff = Dgpu[0*8 + 0*4 + dir];
-
-    for( j=0; j<s->num_blocks; j++ ){
-
-      DD_start = s->block[j].start*l->num_lattice_site_var;
-
-      Dbuff = s->op.D + (DD_start/12)*36;
-
-      ind = index[dir];
-      for ( i=a1; i<n1; i++ ) {
-
-        k = ind[i];
-        //j = neighbor[4*k+T];
-        D_pt = Dbuff + 36*k + 9*dir;
-
-        // copy 9 complex entries: Dgpu_buff <----- D_pt
-        for( h=0; h<9; h++ ){
-          Dgpu_buff[h] = make_cu_cmplx_PRECISION( creal_PRECISION(D_pt[h]), cimag_PRECISION(D_pt[h]) );
-        }
-
-        Dgpu_buff += 9;
-
-      }
-
-    }
-
-    // copy recently created information to the GPU
-    cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).op.Dgpu[0*8 + 0*4 + dir] ), nr_elems_Dgpu[0*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMemcpy((s->s_on_gpu_cpubuff).op.Dgpu[0*8 + 0*4 + dir], Dgpu[0*8 + 0*4 + dir], nr_elems_Dgpu[0*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION), cudaMemcpyHostToDevice) );
-
-  }
-
-  // ... for plus, do ODD
-
-  //amount = _ODD_SITES;
-
-  for( dir=0; dir<4; dir++ ){
-
-    a1 = s->dir_length_even[dir]; n1 = a1 + s->dir_length_odd[dir];
-
-    nr_elems_Dgpu[0*8 + 1*4 + dir] =  (n1-a1);
-    nr_elems_Dgpu[0*8 + 1*4 + dir] *= s->num_blocks;
-
-    // 0*8 + 1*4 + dir ---> PLUS(0), ODD(1), dir
-    Dgpu[0*8 + 1*4 + dir] = (cu_cmplx_PRECISION*) malloc( nr_elems_Dgpu[0*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) );
-
-    Dgpu_buff = Dgpu[0*8 + 1*4 + dir];
-
-    for( j=0; j<s->num_blocks; j++ ){
-
-      DD_start = s->block[j].start*l->num_lattice_site_var;
-
-      Dbuff = s->op.D + (DD_start/12)*36;
-
-      ind = index[dir];
-      for ( i=a1; i<n1; i++ ) {
-
-        k = ind[i];
-        //j = neighbor[4*k+T];
-        D_pt = Dbuff + 36*k + 9*dir;
-
-        // copy 9 complex entries: Dgpu_buff <----- D_pt
-        for( h=0; h<9; h++ ){
-          Dgpu_buff[h] = make_cu_cmplx_PRECISION( creal_PRECISION(D_pt[h]), cimag_PRECISION(D_pt[h]) );
-        }
-
-        Dgpu_buff += 9;
-
-      }
-
-    }
-
-    // copy recently created information to the GPU
-    cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).op.Dgpu[0*8 + 1*4 + dir] ), nr_elems_Dgpu[0*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMemcpy((s->s_on_gpu_cpubuff).op.Dgpu[0*8 + 1*4 + dir], Dgpu[0*8 + 1*4 + dir], nr_elems_Dgpu[0*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION), cudaMemcpyHostToDevice) );
-
-  }
-
-  // then do MINUS
-
-  // ... for minus, do EVEN
-
-  //amount = _EVEN_SITES;
-
-  for( dir=0; dir<4; dir++ ){
-
-    a1 = 0; n1 = s->dir_length_even[dir];
-    a2 = n1; n2 = a2 + s->dir_length_odd[dir];
-
-    nr_elems_Dgpu[1*8 + 0*4 + dir] =  (n2-a2);
-    nr_elems_Dgpu[1*8 + 0*4 + dir] *= s->num_blocks;
-
-    // 1*8 + 0*4 + dir ---> MINUS(1), EVEN(0), dir
-    Dgpu[1*8 + 0*4 + dir] = (cu_cmplx_PRECISION*) malloc( nr_elems_Dgpu[1*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) );
-
-    Dgpu_buff = Dgpu[1*8 + 0*4 + dir];
-
-    for( j=0; j<s->num_blocks; j++ ){
-
-      DD_start = s->block[j].start*l->num_lattice_site_var;
-
-      Dbuff = s->op.D + (DD_start/12)*36;
-
-      ind = index[dir];
-      for ( i=a2; i<n2; i++ ) {
-
-        k = ind[i];
-        //j = neighbor[4*k+T];
-        D_pt = Dbuff + 36*k + 9*dir;
-
-        // copy 9 complex entries: Dgpu_buff <----- D_pt
-        for( h=0; h<9; h++ ){
-          Dgpu_buff[h] = make_cu_cmplx_PRECISION( creal_PRECISION(D_pt[h]), cimag_PRECISION(D_pt[h]) );
-        }
-
-        Dgpu_buff += 9;
-
-      }
-
-    }
-
-
-
-    // TODO: remove
-    /*
-    Dgpu_buff = Dgpu[1*8 + 0*4 + dir];
-    j=11;
-    Dgpu_buff += (n2-a2)*9*11;
-    DD_start = s->block[j].start*l->num_lattice_site_var;
-    Dbuff = s->op.D + (DD_start/12)*36;
-    i = a2;
-    k = s->oe_index[dir][i];
-    D_pt = Dbuff + 36*k + 9*dir;
-    if( dir==1 ){
-      printf( "(DD_start=%d,k=%d,dir=%d) D_pt[0] = %f + i%f, Dgpu_buff[0] = %f + i%f\n", DD_start, k, dir, creal_PRECISION(D_pt[ 0 ]), cimag_PRECISION(D_pt[ 0 ]), \
-                                                                      cu_creal_PRECISION(Dgpu_buff[ 0 ]), cu_cimag_PRECISION(Dgpu_buff[ 0 ]) );
-    }
-    */
-
-
-
-    // copy recently created information to the GPU
-    cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).op.Dgpu[1*8 + 0*4 + dir] ), nr_elems_Dgpu[1*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMemcpy((s->s_on_gpu_cpubuff).op.Dgpu[1*8 + 0*4 + dir], Dgpu[1*8 + 0*4 + dir], nr_elems_Dgpu[1*8 + 0*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION), cudaMemcpyHostToDevice) );
-
-  }
-
-  // ... for minus, do ODD
-
-  //amount = _ODD_SITES;
-
-  for( dir=0; dir<4; dir++ ){
-
-    a1 = s->dir_length_even[dir]; n1 = a1 + s->dir_length_odd[dir];
-    a2 = 0; n2 = a1;
-
-    nr_elems_Dgpu[1*8 + 1*4 + dir] =  (n2-a2);
-    nr_elems_Dgpu[1*8 + 1*4 + dir] *= s->num_blocks;
-
-    // 1*8 + 1*4 + dir ---> MINUS(1), ODD(1), dir
-    Dgpu[1*8 + 1*4 + dir] = (cu_cmplx_PRECISION*) malloc( nr_elems_Dgpu[1*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) );
-
-    Dgpu_buff = Dgpu[1*8 + 1*4 + dir];
-
-    for( j=0; j<s->num_blocks; j++ ){
-
-      DD_start = s->block[j].start*l->num_lattice_site_var;
-
-      Dbuff = s->op.D + (DD_start/12)*36;
-
-      ind = index[dir];
-      for ( i=a2; i<n2; i++ ) {
-
-        k = ind[i];
-        //j = neighbor[4*k+T];
-        D_pt = Dbuff + 36*k + 9*dir;
-
-        // copy 9 complex entries: Dgpu_buff <----- D_pt
-        for( h=0; h<9; h++ ){
-          Dgpu_buff[h] = make_cu_cmplx_PRECISION( creal_PRECISION(D_pt[h]), cimag_PRECISION(D_pt[h]) );
-        }
-
-        Dgpu_buff += 9;
-
-      }
-
-    }
-
-    // copy recently created information to the GPU
-    cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).op.Dgpu[1*8 + 1*4 + dir] ), nr_elems_Dgpu[1*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION) ) );
-    cuda_safe_call( cudaMemcpy((s->s_on_gpu_cpubuff).op.Dgpu[1*8 + 1*4 + dir], Dgpu[1*8 + 1*4 + dir], nr_elems_Dgpu[1*8 + 1*4 + dir] * 9 * sizeof(cu_cmplx_PRECISION), cudaMemcpyHostToDevice) );
-
-  }
-
-  // copying nr_elems_Dgpu to the GPU
-  memcpy( (s->s_on_gpu_cpubuff).op.nr_elems_Dgpu, nr_elems_Dgpu, 16*sizeof(int) );
-  //cuda_safe_call( cudaMemcpy( (s->s_on_gpu_cpubuff).op.nr_elems_Dgpu, nr_elems_Dgpu, 16*sizeof(int), cudaMemcpyHostToDevice ) );
-
-  //----------------------------------------------------------------
-
-  // TODO: remove the following lines for setting GAMMA-related data, and rather put constants within hopping CUDA kernels
-  cu_cmplx_PRECISION *gamma_info_vals_buff = (s->s_on_gpu_cpubuff).gamma_info_vals;
-  int *gamma_info_coo_buff = (s->s_on_gpu_cpubuff).gamma_info_coo;
-  /* even spins - T direction */
-  gamma_info_vals_buff[0] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_T_SPIN0_VAL), cimag_PRECISION(GAMMA_T_SPIN0_VAL));
-  gamma_info_vals_buff[1] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_T_SPIN2_VAL), cimag_PRECISION(GAMMA_T_SPIN2_VAL));
-  /* odd spins */
-  gamma_info_vals_buff[2] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_T_SPIN1_VAL), cimag_PRECISION(GAMMA_T_SPIN1_VAL));
-  gamma_info_vals_buff[3] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_T_SPIN3_VAL), cimag_PRECISION(GAMMA_T_SPIN3_VAL));
-  /* and for the other directions */
-  /* Z */
-  gamma_info_vals_buff[4] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Z_SPIN0_VAL), cimag_PRECISION(GAMMA_Z_SPIN0_VAL));
-  gamma_info_vals_buff[5] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Z_SPIN2_VAL), cimag_PRECISION(GAMMA_Z_SPIN2_VAL));
-  gamma_info_vals_buff[6] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Z_SPIN1_VAL), cimag_PRECISION(GAMMA_Z_SPIN1_VAL));
-  gamma_info_vals_buff[7] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Z_SPIN3_VAL), cimag_PRECISION(GAMMA_Z_SPIN3_VAL));
-  /* Y */
-  gamma_info_vals_buff[8] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Y_SPIN0_VAL), cimag_PRECISION(GAMMA_Y_SPIN0_VAL));
-  gamma_info_vals_buff[9] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Y_SPIN2_VAL), cimag_PRECISION(GAMMA_Y_SPIN2_VAL));
-  gamma_info_vals_buff[10] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Y_SPIN1_VAL), cimag_PRECISION(GAMMA_Y_SPIN1_VAL));
-  gamma_info_vals_buff[11] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_Y_SPIN3_VAL), cimag_PRECISION(GAMMA_Y_SPIN3_VAL));
-  /* X */
-  gamma_info_vals_buff[12] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_X_SPIN0_VAL), cimag_PRECISION(GAMMA_X_SPIN0_VAL));
-  gamma_info_vals_buff[13] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_X_SPIN2_VAL), cimag_PRECISION(GAMMA_X_SPIN2_VAL));
-  gamma_info_vals_buff[14] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_X_SPIN1_VAL), cimag_PRECISION(GAMMA_X_SPIN1_VAL));
-  gamma_info_vals_buff[15] = make_cu_cmplx_PRECISION(creal_PRECISION(GAMMA_X_SPIN3_VAL), cimag_PRECISION(GAMMA_X_SPIN3_VAL));
-  /* even spins - T direction */
-  gamma_info_coo_buff[0] = GAMMA_T_SPIN0_CO;
-  gamma_info_coo_buff[1] = GAMMA_T_SPIN2_CO;
-  /* odd spins */
-  gamma_info_coo_buff[2] = GAMMA_T_SPIN1_CO;
-  gamma_info_coo_buff[3] = GAMMA_T_SPIN3_CO;
-  /* and for the other directions */
-  /* Z */
-  gamma_info_coo_buff[4] = GAMMA_Z_SPIN0_CO;
-  gamma_info_coo_buff[5] = GAMMA_Z_SPIN2_CO;
-  gamma_info_coo_buff[6] = GAMMA_Z_SPIN1_CO;
-  gamma_info_coo_buff[7] = GAMMA_Z_SPIN3_CO;
-  /* Y */
-  gamma_info_coo_buff[8] = GAMMA_Y_SPIN0_CO;
-  gamma_info_coo_buff[9] = GAMMA_Y_SPIN2_CO;
-  gamma_info_coo_buff[10] = GAMMA_Y_SPIN1_CO;
-  gamma_info_coo_buff[11] = GAMMA_Y_SPIN3_CO;
-  /* X */
-  gamma_info_coo_buff[12] = GAMMA_X_SPIN0_CO;
-  gamma_info_coo_buff[13] = GAMMA_X_SPIN2_CO;
-  gamma_info_coo_buff[14] = GAMMA_X_SPIN1_CO;
-  gamma_info_coo_buff[15] = GAMMA_X_SPIN3_CO;
-
-  (s->s_on_gpu_cpubuff).num_block_even_sites = s->num_block_even_sites;
-  (s->s_on_gpu_cpubuff).num_block_odd_sites = s->num_block_odd_sites;
-  (s->s_on_gpu_cpubuff).block_vector_size = s->block_vector_size;
-
-  // The following variable is an indirect indication of the necessary work
-  // over boundaries
-  int *bbl = s->block_boundary_length;
-  s->tot_num_boundary_work = 0;
-  for( dir=0; dir<4; dir++ ){
-
-    // +
-
-    s->num_boundary_sites[dir*2] = (bbl[2*dir+1] - bbl[2*dir])/2;
-    s->tot_num_boundary_work += s->num_boundary_sites[dir*2];
-
-    // -
-
-    s->num_boundary_sites[dir*2+1] = (bbl[2*dir+2] - bbl[2*dir+1])/2;
-    s->tot_num_boundary_work += s->num_boundary_sites[dir*2+1];
-
-  }
-
-  /*
-
-  // computing the nr of boundary sites, in general (for D=4)
-  s->tot_num_boundary_sites = 0;
-
-  // 1. corners: (D-4)-dimensional objects, amount: 2^D = 2^4, neighbors: D
-  s->tot_num_boundary_sites += 16;
-  nr_bound_neighbors[0] = 4;
-
-  // 2. edges: (D-3)-dimensional objects of length 2^1, amount: 32, neighbors: (D-1)
-  s->tot_num_boundary_sites += 32*2;
-  nr_bound_neighbors[0] = 3;
-
-  // 3. faces: (D-2)-dimensional objects of length 2^2, amount: 24, neighbors: (D-2)
-  s->tot_num_boundary_sites += 24*4;
-  nr_bound_neighbors[0] = 2;
-
-  // 4. sides: (D-1)-dimensional objects of length 2^3, amount: 8, neighbors: (D-2)
-  s->tot_num_boundary_sites += 8*8;
-  nr_bound_neighbors[0] = 1;
-
-  */
-
-  // Space to store values from dot products within block solves in SAP
-  cuda_safe_call( cudaMalloc( (void**) &( (s->s_on_gpu_cpubuff).alphas ), s->num_blocks * sizeof(cu_cmplx_PRECISION) ) );
-
-  // After all the allocations and definitions associated to s->s_on_gpu_cpubuff, it's time to move to the GPU
-  cuda_safe_call( cudaMemcpy(s->s_on_gpu, &(s->s_on_gpu_cpubuff), 1*sizeof(schwarz_PRECISION_struct_on_gpu), cudaMemcpyHostToDevice) );
-
-  int mu_dir, inv_mu_dir, mu, boundary_size;
-
-  // TODO: move the following code for allocating memory for boundary_table_gpu to the appropriate place
-
-  for( mu=0; mu<4; mu++ ){
-
-    //-1
-    dir = -1;
-    mu_dir = 2*mu-MIN(dir,0);
-    inv_mu_dir = 2*mu+1+MIN(dir,0);
-    boundary_size = (s->op.c).num_boundary_sites[mu_dir];
-    //cuda_safe_call( cudaMemcpy( (s->op.c).boundary_table_gpu[inv_mu_dir], (s->op.c).boundary_table[inv_mu_dir], boundary_size*sizeof(int), cudaMemcpyHostToDevice ) );
-    cuda_safe_call( cudaMalloc( (void**) &((s->op.c).boundary_table_gpu[inv_mu_dir]), boundary_size*sizeof(int) ) );
-    boundary_size *= l->num_lattice_site_var;
-    cuda_safe_call( cudaMalloc( (void**) &((s->op.c).buffer_gpu[mu_dir]), boundary_size*sizeof(cu_cmplx_PRECISION) ) );
-
-    //+1
-    dir = 1;
-    mu_dir = 2*mu-MIN(dir,0);
-    inv_mu_dir = 2*mu+1+MIN(dir,0);
-    boundary_size = (s->op.c).num_boundary_sites[mu_dir];
-    cuda_safe_call( cudaMalloc( (void**) &((s->op.c).boundary_table_gpu[inv_mu_dir]), boundary_size*sizeof(int) ) );
-    boundary_size *= l->num_lattice_site_var;
-    cuda_safe_call( cudaMalloc( (void**) &((s->op.c).buffer_gpu[mu_dir]), boundary_size*sizeof(cu_cmplx_PRECISION) ) );
-
-  }
-
-  for( mu=0; mu<4; mu++ ){
-
-    //-1
-    dir = -1;
-    mu_dir = 2*mu-MIN(dir,0);
-    inv_mu_dir = 2*mu+1+MIN(dir,0);
-    boundary_size = (s->op.c).num_boundary_sites[mu_dir];
-    cuda_safe_call( cudaMemcpy( (s->op.c).boundary_table_gpu[inv_mu_dir], (s->op.c).boundary_table[inv_mu_dir], boundary_size*sizeof(int), cudaMemcpyHostToDevice ) );
-
-    //+1
-    dir = 1;
-    mu_dir = 2*mu-MIN(dir,0);
-    inv_mu_dir = 2*mu+1+MIN(dir,0);
-    boundary_size = (s->op.c).num_boundary_sites[mu_dir];
-    cuda_safe_call( cudaMemcpy( (s->op.c).boundary_table_gpu[inv_mu_dir], (s->op.c).boundary_table[inv_mu_dir], boundary_size*sizeof(int), cudaMemcpyHostToDevice ) );
-
-  }
-
-}
-#endif
-
-
 void additive_schwarz_PRECISION( vector_PRECISION phi, vector_PRECISION D_phi, vector_PRECISION eta, const int cycles, int res, 
                                  schwarz_PRECISION_struct *s, level_struct *l, struct Thread *threading ) {
   
@@ -2560,262 +1665,6 @@ void schwarz_PRECISION( vector_PRECISION phi, vector_PRECISION D_phi, vector_PRE
 }
 
 
-#ifdef CUDA_OPT
-
-// FIXME: profiling is not measuring correctly
-
-void schwarz_PRECISION_CUDA( vector_PRECISION phi, vector_PRECISION D_phi, vector_PRECISION eta, const int cycles, int res,
-                             schwarz_PRECISION_struct *s, level_struct *l, struct Thread *threading ) {
-  
-  START_NO_HYPERTHREADS(threading)
-
-  int color, k, mu, nb = s->num_blocks, init_res = res;
-
-  // spinors on the GPU to be used on computations
-  cuda_vector_PRECISION r_dev = (s->cu_s).buf1, x_dev = (s->cu_s).buf3, \
-                        latest_iter_dev = (s->cu_s).buf2, Dphi_dev = (s->cu_s).buf4, \
-                        eta_dev = (s->cu_s).buf5, D_phi_dev = (s->cu_s).buf6;
-
-  SYNC_CORES(threading)
-  
-  int nb_thread_start;
-  int nb_thread_end;
-  compute_core_start_end_custom(0, nb, &nb_thread_start, &nb_thread_end, l, threading, 1);
-
-  // Creation of CUDA Events, for time measurement and GPU sync
-  cudaEvent_t start_event_copy, stop_event_copy, start_event_comp, stop_event_comp;
-  cuda_safe_call( cudaEventCreate(&start_event_copy) );
-  cuda_safe_call( cudaEventCreate(&stop_event_copy) );
-  cuda_safe_call( cudaEventCreate(&start_event_comp) );
-  cuda_safe_call( cudaEventCreate(&stop_event_comp) );
-  // TODO: generalization to more than one streams pending
-  cudaStream_t *streams_schwarz = s->streams;
-
-  cuda_vector_PRECISION_copy( (void*)eta_dev, (void*)eta, nb_thread_start*s->block_vector_size, \
-                              (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _H2D, _CUDA_SYNC, 0, streams_schwarz );
-
-  if ( res == _NO_RES ) {
-    cuda_vector_PRECISION_copy( (void*)r_dev, (void*)eta_dev, nb_thread_start*s->block_vector_size,
-                                (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _D2D, _CUDA_SYNC, 0, streams_schwarz );
-    cuda_vector_PRECISION_define( x_dev, make_cu_cmplx_PRECISION(0,0), nb_thread_start*s->block_vector_size, \
-                                  (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _CUDA_SYNC, 0, streams_schwarz );
-  } else {
-    cuda_vector_PRECISION_copy( (void*)x_dev, (void*)phi, nb_thread_start*s->block_vector_size,
-                                (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _H2D, _CUDA_SYNC, 0, streams_schwarz );
-  }
-
-  START_MASTER(threading)
-  if ( res == _NO_RES ) {
-    // setting boundary values to zero
-    cuda_vector_PRECISION_define( x_dev, make_cu_cmplx_PRECISION(0,0), l->inner_vector_size, \
-                                  (l->schwarz_vector_size - l->inner_vector_size), l, _CUDA_SYNC, 0, streams_schwarz );
-  }
-  END_MASTER(threading)
-
-  SYNC_CORES(threading)
-
-  for ( k=0; k<cycles; k++ ) {
-    for ( color=0; color<s->num_colors; color++ ) {
-      if ( res == _RES ) {
-        START_LOCKED_MASTER(threading)
-        for ( mu=0; mu<4; mu++ ) {
-          cuda_ghost_update_PRECISION( (k==0 && init_res == _RES)?x_dev:latest_iter_dev, mu, +1, &(s->op.c), l );
-          cuda_ghost_update_PRECISION( (k==0 && init_res == _RES)?x_dev:latest_iter_dev, mu, -1, &(s->op.c), l );
-        }
-        END_LOCKED_MASTER(threading)
-      } else {
-        // we need a barrier between black and white blocks
-        SYNC_CORES(threading)
-      }
-
-      // TODO: add profiling !
-
-      if ( res == _RES ) {
-        if ( k==0 && init_res == _RES ) {
-          cuda_block_d_plus_clover_PRECISION( Dphi_dev, x_dev, s->nr_DD_blocks_notin_comms[color], s, l, \
-                                              no_threading, 0, streams_schwarz, color, \
-                                              (s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color] );
-          cuda_block_PRECISION_boundary_op( Dphi_dev, x_dev, s->nr_DD_blocks_notin_comms[color], s, l, \
-                                            no_threading, 0, streams_schwarz, color, \
-                                            (s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color] );
-          cuda_block_vector_PRECISION_minus( r_dev, eta_dev, Dphi_dev, s->nr_DD_blocks_notin_comms[color], s, l, \
-                                             no_threading, 0, streams_schwarz, color, \
-                                             (s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color] );
-        } else {
-          cuda_n_block_PRECISION_boundary_op( r_dev, latest_iter_dev, s->nr_DD_blocks_notin_comms[color], s, l, \
-                                              no_threading, 0, streams_schwarz, color, \
-                                              (s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color] );
-        }
-      }
-
-      // local minres updates x, r and latest iter
-      cuda_block_solve_oddeven_PRECISION( (cuda_vector_PRECISION)x_dev, (cuda_vector_PRECISION)r_dev, (cuda_vector_PRECISION)latest_iter_dev,
-                                          0, s->nr_DD_blocks_notin_comms[color], s, l, no_threading, 0, streams_schwarz, 0, color,
-                                          (s->cu_s).DD_blocks_notin_comms[color], s->DD_blocks_notin_comms[color] );
-
-      if ( res == _RES ) {
-        START_LOCKED_MASTER(threading)
-        for ( mu=0; mu<4; mu++ ) {
-          cuda_ghost_update_wait_PRECISION( (k==0 && init_res == _RES)?x_dev:latest_iter_dev, mu, +1, &(s->op.c), l );
-          cuda_ghost_update_wait_PRECISION( (k==0 && init_res == _RES)?x_dev:latest_iter_dev, mu, -1, &(s->op.c), l );
-          // TODO: remove the following line?
-          cuda_safe_call( cudaDeviceSynchronize() );
-        }
-        END_LOCKED_MASTER(threading)
-      } else {
-        // we need a barrier between black and white blocks
-        SYNC_CORES(threading)
-        // TODO: is this sync really needed .. ?
-        cuda_safe_call( cudaDeviceSynchronize() );
-      }
-
-      // TODO: remove the following line .. ?
-      cuda_safe_call( cudaDeviceSynchronize() );
-
-      if ( res == _RES ) {
-        if ( k==0 && init_res == _RES ) {
-
-          cuda_block_d_plus_clover_PRECISION( Dphi_dev, x_dev, s->nr_DD_blocks_in_comms[color], s, l, no_threading, 0, \
-                                              streams_schwarz, color, (s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color] );
-          cuda_block_PRECISION_boundary_op( Dphi_dev, x_dev, s->nr_DD_blocks_in_comms[color], s, l, no_threading, 0, \
-                                            streams_schwarz, color, (s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color] );
-          cuda_block_vector_PRECISION_minus( r_dev, eta_dev, Dphi_dev, s->nr_DD_blocks_in_comms[color], s, l, no_threading, 0, \
-                                             streams_schwarz, color, (s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color] );
-        } else {
-          cuda_n_block_PRECISION_boundary_op( r_dev, latest_iter_dev, s->nr_DD_blocks_in_comms[color], s, l, no_threading, 0, \
-                                              streams_schwarz, color, (s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color] );
-        }
-      }
-
-      // local minres updates x, r and latest iter
-      cuda_block_solve_oddeven_PRECISION( (cuda_vector_PRECISION)x_dev, (cuda_vector_PRECISION)r_dev, (cuda_vector_PRECISION)latest_iter_dev,
-                                          0, s->nr_DD_blocks_in_comms[color], s, l, no_threading, 0, streams_schwarz, 0, color,
-                                          (s->cu_s).DD_blocks_in_comms[color], s->DD_blocks_in_comms[color] );
-
-      res = _RES;
-    }
-  }
-
-  // TODO: remove the following line?
-  cuda_safe_call( cudaDeviceSynchronize() );
-
-  if ( l->relax_fac != 1.0 ){
-    cuda_vector_PRECISION_scale( x_dev, make_cu_cmplx_PRECISION( l->relax_fac, 0 ), nb_thread_start*s->block_vector_size, \
-                                 (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _CUDA_SYNC, 0, streams_schwarz );
-    cuda_vector_PRECISION_copy( (void*)phi, (void*)x_dev, nb_thread_start*s->block_vector_size,
-                                (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _D2H, _CUDA_SYNC, 0, streams_schwarz );
-  } else {
-    cuda_vector_PRECISION_copy( (void*)phi, (void*)x_dev, nb_thread_start*s->block_vector_size,
-                                (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _D2H, _CUDA_SYNC, 0, streams_schwarz );
-  }
-
-  // TODO: remove the following line?
-  cuda_safe_call( cudaDeviceSynchronize() );
-
-  //gettimeofday(&start, NULL);
-
-  if( D_phi != NULL ){
-
-    START_LOCKED_MASTER(threading)
-    for ( mu=0; mu<4; mu++ ) {
-      cuda_ghost_update_PRECISION( latest_iter_dev, mu, +1, &(s->op.c), l );
-      cuda_ghost_update_PRECISION( latest_iter_dev, mu, -1, &(s->op.c), l );
-    }
-    END_LOCKED_MASTER(threading)
-
-    // 1. for all those DD blocks of color=0 and NOT INVOLVED in comms: n_boundary_op( ... ), vector_PRECISION_minus( ... ), vector_PRECISION_scale( ... )
-
-    cuda_n_block_PRECISION_boundary_op( r_dev, latest_iter_dev, s->nr_DD_blocks_notin_comms[0], s, l, no_threading, 0, \
-                                        streams_schwarz, 0, (s->cu_s).DD_blocks_notin_comms[0], s->DD_blocks_notin_comms[0] );
-    cuda_block_vector_PRECISION_minus( D_phi_dev, eta_dev, r_dev, s->nr_DD_blocks_notin_comms[0], s, l, no_threading, 0, \
-                                       streams_schwarz, 0, (s->cu_s).DD_blocks_notin_comms[0], s->DD_blocks_notin_comms[0] );
-    if( l->relax_fac != 1.0 ){
-      cuda_block_vector_PRECISION_scale( D_phi_dev, D_phi_dev, make_cu_cmplx_PRECISION( l->relax_fac, 0 ), s->nr_DD_blocks_notin_comms[0], s, l, no_threading, 0,
-                                         streams_schwarz, 0, (s->cu_s).DD_blocks_notin_comms[0], s->DD_blocks_notin_comms[0] );
-    }
-
-    // 2. for all those DD blocks of color=1: vector_PRECISION_minus( ... ), vector_PRECISION_scale( ... )
-
-    cuda_block_vector_PRECISION_minus( D_phi_dev, eta_dev, r_dev, s->nr_DD_blocks[1], s, l, no_threading, 0, \
-                                       streams_schwarz, 1, (s->cu_s).DD_blocks[1], s->DD_blocks[1] );
-
-    if( l->relax_fac != 1.0 ){
-      cuda_block_vector_PRECISION_scale( D_phi_dev, D_phi_dev, make_cu_cmplx_PRECISION( l->relax_fac, 0 ), s->nr_DD_blocks[1], s, l, no_threading, 0,
-                                         streams_schwarz, 1, (s->cu_s).DD_blocks[1], s->DD_blocks[1] );
-    }
-
-    START_LOCKED_MASTER(threading)
-    for ( mu=0; mu<4; mu++ ) {
-      cuda_ghost_update_wait_PRECISION( latest_iter_dev, mu, +1, &(s->op.c), l );
-      cuda_ghost_update_wait_PRECISION( latest_iter_dev, mu, -1, &(s->op.c), l );
-    }
-    END_LOCKED_MASTER(threading)
-
-    // 3. for all those DD blocks of color=0 and INVOLVED in comms: n_boundary_op( ... ), vector_PRECISION_minus( ... ), vector_PRECISION_scale( ... )
-
-    cuda_n_block_PRECISION_boundary_op( r_dev, latest_iter_dev, s->nr_DD_blocks_in_comms[0], s, l, no_threading, 0, \
-                                        streams_schwarz, 0, (s->cu_s).DD_blocks_in_comms[0], s->DD_blocks_in_comms[0] );
-    cuda_block_vector_PRECISION_minus( D_phi_dev, eta_dev, r_dev, s->nr_DD_blocks_in_comms[0], s, l, no_threading, 0, \
-                                       streams_schwarz, 0, (s->cu_s).DD_blocks_in_comms[0], s->DD_blocks_in_comms[0] );
-    if( l->relax_fac != 1.0 ){
-      cuda_block_vector_PRECISION_scale( D_phi_dev, D_phi_dev, make_cu_cmplx_PRECISION( l->relax_fac, 0 ), s->nr_DD_blocks_in_comms[0], s, l, no_threading, 0,
-                                         streams_schwarz, 0, (s->cu_s).DD_blocks_in_comms[0], s->DD_blocks_in_comms[0] );
-    }
-
-    // TODO: remove the following line?
-    cuda_safe_call( cudaDeviceSynchronize() );
-
-  }
-
-  // TODO: remove the following line?
-  cuda_safe_call( cudaDeviceSynchronize() );
-
-  if( D_phi != NULL ){
-    cuda_vector_PRECISION_copy( (void*)D_phi, (void*)D_phi_dev, nb_thread_start*s->block_vector_size, \
-                                (nb_thread_end-nb_thread_start)*s->block_vector_size, l, _D2H, _CUDA_SYNC, 0, streams_schwarz );
-  }
-
-  // TODO: remove the following line?
-  cuda_safe_call( cudaDeviceSynchronize() );
-
-#ifdef SCHWARZ_RES
-  // TODO: portion pending to port to GPU/CUDA
-  START_LOCKED_MASTER(threading)
-  if ( D_phi == NULL ) {
-    for ( mu=0; mu<4; mu++ ) {
-      ghost_update_PRECISION( latest_iter, mu, +1, &(s->op.c), l );
-      ghost_update_PRECISION( latest_iter, mu, -1, &(s->op.c), l );
-    }
-    
-    for ( i=0; i<nb; i++ ) {
-      if ( s->block[i].no_comm ) {
-        n_boundary_op( r, latest_iter, i, s, l );
-      }
-    }
-    
-    for ( mu=0; mu<4; mu++ ) {
-      ghost_update_wait_PRECISION( latest_iter, mu, +1, &(s->op.c), l );
-      ghost_update_wait_PRECISION( latest_iter, mu, -1, &(s->op.c), l );
-    }
-    
-    for ( i=0; i<nb; i++ ) {
-      if ( !s->block[i].no_comm ) {
-        n_boundary_op( r, latest_iter, i, s, l );
-      }
-    }
-  }
-  double rnorm = global_norm_PRECISION( r, 0, l->inner_vector_size, l, no_threading );
-  char number[3]; sprintf( number, "%2d", 31+l->depth ); printf0("\033[1;%2sm|", number );
-  printf0(" ---- depth: %d, c: %d, schwarz iter %2d, norm: %11.6le |", l->depth, s->num_colors, k, rnorm );
-  printf0("\033[0m\n"); fflush(0);
-  END_LOCKED_MASTER(threading)
-#endif
-
-  END_NO_HYPERTHREADS(threading)
-}
-#endif
-
-
 void sixteen_color_schwarz_PRECISION( vector_PRECISION phi, vector_PRECISION D_phi, vector_PRECISION eta, const int cycles, int res, 
                                       schwarz_PRECISION_struct *s, level_struct *l, struct Thread *threading ) {
   
@@ -3019,18 +1868,6 @@ void schwarz_PRECISION_def( schwarz_PRECISION_struct *s, operator_double_struct 
   schwarz_layout_PRECISION_define( s, l );
   schwarz_PRECISION_setup( s, op, l );
 }
-
-
-#ifdef CUDA_OPT
-void schwarz_PRECISION_def_CUDA( schwarz_PRECISION_struct *s, operator_double_struct *op, level_struct *l ) {
-
-  if( l->depth==0 && g.odd_even ){
-    schwarz_PRECISION_alloc_CUDA( s, l );
-    schwarz_PRECISION_setup_CUDA( s, op, l );
-  }
-
-}
-#endif
 
 
 void schwarz_PRECISION_mvm_testfun( schwarz_PRECISION_struct *s, level_struct *l, struct Thread *threading ) {
